♻️ 関数をファイル分けした
