+++ Makefile
@@ -6,7 +6,7 @@
 #    By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/08 01:21:55 by miyuu             #+#    #+#              #
-#    Updated: 2025/05/24 13:53:13 by miyuu            ###   ########.fr        #
+#    Updated: 2025/05/25 00:21:39 by miyuu            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -32,7 +32,9 @@ LOAD_SRCS = load/init_cubdata.c \
 			load/free_data.c \
 			load/utils/ft_str_lst.c \
 			load/utils/free_str_array.c \
-			load/utils/error_print_exit.c
+			load/utils/error_print_exit.c \
+			load/utils/is_empty_line.c \
+			load/utils/parse_utils.c
 
 #debugディレクトリは最終的に削除する
 DEBUG_SRCS = debug/debug_print_data.c \
